refactor(#1401): centralize file event contracts
