+++ .workflow/records/1407-adr-045-manager-1407-drift.json
@@ -67,6 +67,17 @@
         "src/scistudio/api/routes/workflows.py"
       ],
       "reason": "Keep ADR-045 source_id contract consistent by allowing workflow write bodies as well as existing headers to carry source_id."
+    },
+    {
+      "approved_by": null,
+      "exclude": [],
+      "include": [
+        "src/scistudio/api/file_contracts.py",
+        "src/scistudio/api/routes/projects.py",
+        "src/scistudio/api/routes/workflow_watcher.py",
+        "src/scistudio/api/ws.py"
+      ],
+      "reason": "Resolve Sentrux max_cycles by moving shared file event contract constants out of route modules into a neutral API contract module."
     }
   ],
   "branch": "track/adr-045/version-vector",
@@ -122,6 +133,15 @@
       "exit_code": 0,
       "name": "full_audit",
       "output_path": "docs/audit/adr-045-full-audit.json",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "pytest tests/api/test_workflow_version_vector.py tests/api/test_file_version_vector.py tests/api/test_workflow_changed_event_schema.py tests/api/routes/test_workflow_watcher_fallback.py tests/api/test_reload_on_save.py tests/integration/test_race_autosave.py tests/integration/test_race_lineage_restore.py tests/integration/test_race_agent_write.py tests/integration/test_race_external_editor.py tests/integration/test_race_multi_session.py --timeout=60 --no-cov",
+      "exit_code": 0,
+      "name": "adr045_contract_cycle_fix_pytest",
+      "output_path": "22 passed, 1 deprecation warning",
       "status": "pass",
       "summary": {},
       "timestamp": null
