+++ docs/planning/adr-045-version-vector-checklist.md
@@ -211,6 +211,7 @@ Append only.
 | 2026-05-22 | manager | #1407: ADR-041/043 implemented code existed but ADR/spec phase/status remained planning/Draft, hiding doc/signature drift. | Manager opened drift-fix track before ADR-045 implementation dispatch. | #1407 |
 | 2026-05-22 | manager | Owner emphasized contract consistency for ADR-045 dispatch. | Added contract-consistency requirements to spec, checklist, and all A1-A4 dispatch prompts. | #1401 |
 | 2026-05-22 | A3 / manager | A3 found `file.changed` was emitted to EventBus but not forwarded over `/ws`, and file-tab external-editor watcher fallback was absent. | Manager expanded gate scope, added `file.changed` WebSocket forwarding, project-file watcher fallback, workflow body/header `source_id` parity, and ADR race-matrix integration tests. | #1401 |
+| 2026-05-22 | manager | Umbrella CI semantic duplication and A3 Sentrux cycle checks flagged route-level duplication/cycles in the file-event contract wiring. | Manager refactored watcher event flow and moved shared file event constants into `scistudio.api.file_contracts` so route modules no longer import each other for `file.changed` contract truth. | #1401 |
 
 ## 11. Final Readiness
 
@@ -233,6 +234,7 @@ assigned track rows above.
 | 2026-05-22 | A2 / Linnaeus | `019e4dd1-2401-7ec0-9752-d5849a342f5b` | `[x]` | PR #1425 merged to `track/adr-045/version-vector`; no Codex auto-review comments. |
 | 2026-05-22 | A3 / Godel | `019e4dd1-24fb-7d43-af13-37dd74fd67f1` | `[!]` | Stopped on backend contract blocker: `file.changed` not in `/ws` outbound allowlist; manager accepted finding. |
 | 2026-05-22 | manager | local | `[~]` | Added backend contract-consistency fix and five ADR race-matrix integration test files; targeted backend regression suite passed 22 tests. |
+| 2026-05-22 | manager | local | `[~]` | Fixed semantic duplication ratchet and route-import cycle risk by sharing watcher event flow and moving `file.changed`/allowlist constants to `src/scistudio/api/file_contracts.py`; targeted backend regression suite passed 22 tests. |
 | 2026-05-22 | A4 / audit | pending | `[ ]` | Held until A3 PR/commit exists; prompt prepared at `docs/planning/dispatch-prompts/adr-045-a4-audit-with-context.md`. |
 
 ## 13. Codex Auto-Audit Triage
